requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
tiktoken>=0.5.0

# LangChain (modern API, >=0.2)
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.7

# Vector Store & Embeddings
faiss-cpu>=1.8.0
numpy>=1.24.0

# Data Processing
pandas>=2.0.0
datasets>=2.12.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Optional: For Jupyter notebooks (if needed)
# jupyter>=1.0.0
# ipython>=8.0.0

# Streamlit UI for interactive exploration and dashboards
streamlit>=1.20.0

# Notes:
# - LangChain 0.2+ uses modern invoke() API, not deprecated .run()
# - FAISS CPU version for development; use faiss-gpu for production
# - All dependencies pinned to stable versions
# - No deprecated APIs are used in this project

